Removed check for timeout in Connection since it can cause missing cycles.
Fixes #1
The call to waitMessage() can cause agents to miss cycles. If we
properly want to be safe from a hanging client, we have to check this in
an extra thread. Since this is not a trivial change, I removed this
feature and will introduce a new issue for it that should be fixed after
RoboCup 2015.
Corrected wrong command line parameters in readme